Refatorado tabelas de classes de pessoa
